Clase Television terminada y comprobada 2

+++ src/main/java/Ej3_Electrodomesticos/Lavadora.java
@@ -3,18 +3,18 @@ package Ej3_Electrodomesticos;
 public class Lavadora extends Electrodomestico {
     private int carga;
     private final static int CARGA_DEF=5;
-    private int precioFin;
+    private double precioFin;
     
     public Lavadora(){
         this(PRECIO_DEF, PESO_DEF, CONENE_DEF, COLOR_DEF, CARGA_DEF);
     }
     
-    public Lavadora (int precio_base, int peso){
+    public Lavadora (double precio_base, int peso){
        this(precio_base, peso, CONENE_DEF, COLOR_DEF, CARGA_DEF);
        
     }
     
-    public Lavadora (int precio_base, int peso, char consumo_energetico, String color, int carga){
+    public Lavadora (double precio_base, int peso, char consumo_energetico, String color, int carga){
        super(precio_base, peso, consumo_energetico, color);
        this.carga=carga;
     }
@@ -24,7 +24,7 @@ public class Lavadora extends Electrodomestico {
         
     } 
     
-    public int precioLavadora(Electrodomestico e){
+    public double precioLavadora(Electrodomestico e){
         
         if (carga<30){
             precioFin=e.precioFinal();
